added funtionality to unistall app both from ui and localstorage.

+++ src/Utilites/LocalStorage.js
@@ -1,3 +1,5 @@
+import { Apple } from "lucide-react";
+
 // get the list from localstorage
 export const LoadAppList = () => {
   try {
@@ -10,13 +12,24 @@ export const LoadAppList = () => {
 };
 
 // Save in local storage
-export const updateList = (app) => {
+export const updateList = (apps) => {
   const AppList = LoadAppList();
 
   try {
-    const updateAppList = [...AppList, app];
+    const isDuplicate = AppList.some((p) => p.id === apps.id);
+    if (isDuplicate) return alert("Already Installed");
+    const updateAppList = [...AppList, apps];
     localStorage.setItem("AppList", JSON.stringify(updateAppList));
-    // console.log(updateAppList);
+  } catch (error) {
+    console.log(error);
+  }
+};
+// Remove from AppList
+export const removeFromAppList = (id) => {
+  const AppList = LoadAppList();
+  try {
+    const deleteAppList = AppList.filter((p) => p.id !== id);
+    localStorage.setItem("AppList", JSON.stringify(deleteAppList));
   } catch (error) {
     console.log(error);
   }
